Download necessary software
Added downloads for Librewolf and Revo Uninstaller.

Librewolf is currently the most privacy protected browser and a web-browser is needed for when edge is removed.

Revo Uninstaller will be downloaded so the user can easily remove programs from their computer in case any get added in future windows updates and the user does not want to keep this program.